migrate TestIncrementProposalNumber to simap
